Create logger object from env
When not using Laravel and wanting to pass configuration in via environment variables (e.g. running in docker), there is no way currently to create a logger object to assign to the `salesforce.logger` configuration option. This code will create an object if the config option exists and is a string.